requirements.txt
streamlit
requests
langchain_openai
langchain_community
pandas
langchain
langchain-core
langchain-experimental
langchain-text-splitters
langdetect
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
unstructured
unstructured-client
networkx
pydantic
scikit-learn
azure-cosmos
azure-storage-blob
unstructured[pdf]
opentelemetry-sdk
opentelemetry-instrumentation-langchain
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry
azure-identity
torch
azure-ai-projects
pypdf
docx2txt
unstructured
openpyxl
databricks-sdk
